Remove EventDTO and refactor the Message view model
* Merge ChatRoomEventNum into MessageTypeEnum

* Major refactoring and improvement of MessageTypeEnum

* Adjust MessageSent to support system-generated event-messages

* Make PushNotificationDTO to use the more  generic MessageSentDTO instead

+++ XamApp/XamApp.Android/Notifications/Notifications.cs
@@ -63,7 +63,7 @@ namespace XamApp.Droid.Notifications
             notificationBuilder.SetContentIntent(resultPendingIntent);
 
             var notificationManager = (NotificationManager)context.GetSystemService(Context.NotificationService);
-            notificationManager.Notify(notification.Tag, notification.Id, notificationBuilder.Build());
+            notificationManager.Notify(notification.Tag, (int)notification.Id, notificationBuilder.Build());
         }
 
         public static void ProcessIntent(Intent intent, App.NotificationSource source)
